Restructure into src.
