+++ package.json
@@ -1,8 +1,12 @@
 {
   "name": "cw-smoke",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "src/index.js",
   "license": "MIT",
+  "scripts": {
+    "start": "node src/index.js",
+    "schedule": "SCHEDULE_CRON=true node src/index.js"
+  },
   "dependencies": {
     "@polkadot/api": "1.18.1",
     "@polkadot/util": "2.13.1",
